feat: 유사 견적 페이지 Alert enum
