+++ iOS/Catalog/Catalog/QuotationComplete/QuotationComplete/QuotationCompleteViewAlert.swift
@@ -1,0 +1,21 @@
+//
+//  QuotationCompleteViewAlert.swift
+//  Catalog
+//
+//  Created by 이수민 on 2023/08/23.
+//
+
+import Foundation
+import SwiftUI
+
+
+extension QuotationCompleteView {
+  enum AlertCase {
+    case modify(title: String)
+    case delete
+  }
+}
+
+extension QuotationCompleteView {
+  
+}
