+++ iOS/Catalog/Catalog/QuotationComplete/View/QuotationCompleteSheet.swift
@@ -75,6 +75,8 @@ extension QuotationCompleteSheet: View {
         }
       }
     }
-    
+    .CLDialogFullScreenCover(show: $showAlert, title: String) {
+      
+    }
   }
 }
