+++ iOS/Catalog/Catalog/SimilarQuotation/Intent/SimilarQuotationIntent.swift
@@ -66,7 +66,8 @@ extension SimilarQuotationIntent: SimilarQuotationIntentType, IntentType {
         print("추가하기 버튼 클릭")
         state.alertCase = .addOption
         Quotation.shared.send(action: .similarOptionsAdded(option: state.selectedOptions))
-        
+      case .onTapHelpButton:
+        state.alertCase = .help
       case .optionSelected(let selectedOption):
         if state.selectedOptions.contains(selectedOption) {
           state.selectedOptions = state.selectedOptions.filter { $0 != selectedOption }
